use `unwrap()` to work with Rust standard library locks
Explanation:

`unwrap()` on `RwLock` or `Mutex` can return an `Err` only if there was a `panic!()` during critical section - a time while `MutexGuard` variable is in scope.

In our both cases we never call methods that can panic (f.e. `serde_json::from_str` doesn't panic - it returns a `Result` object instead, and we use `?` on that, not `unwrap()`, so we're good).

In general, you should try to design your code in such a way that critical sections do not produce panics and thus you can safely `unwrap()` your locks. This is exactly what you do here.