Manager task shutdown logic.
Tokio's `oneshot` channel can be used in both synchronous and asynchronous context. In particular its ability to do `send` synchronously makes it very convenient to use in `Drop::drop` method.

We never **wait** for the message to come. Instead we use `try_recv` to check if the shutdown message is there to avoid blocking or awaiting.